Sleep: fixed negeative glitch
If sleep was called 10 seconds or less before the wakeup point, it would raise an error, since our sleep time is subtracted a few seconds to accomodate downloading and processing data.  Now we have added an if statement that checks if there is 15 seconds or less until the wakeup time we just sleep for 5 seconds.